Rename tests to follow conventions
