+++ test/datc/test_datcH.py
@@ -6,7 +6,7 @@ from diplomacy.persistence.order import (
     Support,
 )
 from diplomacy.persistence.unit import UnitType
-from tests.utils import BoardBuilder
+from test.utils import BoardBuilder
 
 
 # These tests are based off https://webdiplomacy.net/doc/DATC_v3_0.html, with
