--- tests/__init__.py
@@ -1,1 +1,0 @@
-    # TODO coring tests?
